[Visualizations] Simplify visualization telemetry (#77145)
* Move visualizations telemetry into visualizations plugin

* Remove x-pack oss_telemetry plugin

* Add unit tests

* Update docs

* Revert not related changes

# Conflicts:
#	x-pack/plugins/oss_telemetry/server/test_utils/index.ts
