--- x-pack/plugins/oss_telemetry/server/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { PluginInitializerContext } from 'src/core/server';
-import { OssTelemetryPlugin } from './plugin';
-
-export const plugin = (context: PluginInitializerContext) => new OssTelemetryPlugin(context);
-
-export * from './plugin';
